requirements.txt
Django
djangorestframework
boto3
gunicorn
psycopg2-binary
whitenoise
environs[django]
dj-rest-auth